requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
joblib==1.3.2
apscheduler==3.10.4
python-dotenv==1.0.0
lightgbm==4.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
